ref: replace V2 with V3 code
