--- test/resources/private.pem
@@ -1,15 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIICWwIBAAKBgQDes/vmu1GHFotfMb9bK7sL8bnqZ5GJuY2M2dBQFOuBaYqPZeJ4
-Z/F7pXXSHey9fKDJ2mu5slutZNlqf+2a+uh+1b4YCFHkJ3fa4UIW3CgXctbav5kR
-hmcmUURldPcjFdKvL5UWaz/YabQEteVEy0Ff5Odfb590LMcUmruXjf3xwwIDAQAB
-AoGAWg3DzfBvLkYSg6L1MyPC1jigKkZJFr0enRcy4j8jYEhSTFwxsOivQIWO2DFb
-mt/7kMQvDX8vDTWbqTIj3aiaPHaM44GpoOr99NMjvLtlrNXNU6YNeDaq6HC7apXr
-pgZJtaK/bVU2xtrMnqF/YZx7vOATB8s1pu2qtcqBybvgWckCQQDvNXFlXEJl2Knw
-gCdgtGC7gDlHdBAxAmEtvKX+l6XUY9C2AkCHnJz2nh8lyR0H1c5OB5Y27zew3t97
-mYCVGZ9PAkEA7lXvcFIGNue/0h7Hcyjng5GevdSrgz+gU/URM8Fa+4Q36ZwbvAzB
-7kF6gnDh2uts2C+5UZjn775neidkCIzJTQJARZn6RbDjjPjO8bCIRKYXY1c602Sc
-9XWoywMSFZMQaE0mxF7BYfQa49M8YuUysj3uyluAxLk0Lh2bTz5p4HhVFwJARil8
-NGehNragrJ2s/4TSKtRKk/2whBHOBAlICCFqxFJXF2mCbKigyp/yDcZnZsgZocbi
-2leWJ+i4k6hW9YFknQJAO647KoL9V96WVnLLt2soEuzrgy4kzSmfmfHjI2Te5h62
-ZoIrnE1KpWjQRbtmkNZHyK/BfYz96ReFnECtJIrruw==
------END RSA PRIVATE KEY-----
