--- test/resources/public.pem
@@ -1,6 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDes/vmu1GHFotfMb9bK7sL8bnq
-Z5GJuY2M2dBQFOuBaYqPZeJ4Z/F7pXXSHey9fKDJ2mu5slutZNlqf+2a+uh+1b4Y
-CFHkJ3fa4UIW3CgXctbav5kRhmcmUURldPcjFdKvL5UWaz/YabQEteVEy0Ff5Odf
-b590LMcUmruXjf3xwwIDAQAB
------END PUBLIC KEY-----
