--- test/resources/response.txt
@@ -1,1 +1,0 @@
-{"edulevel":{"lastupdated":"2017-12-19","source":"2","classification":"C","value":"5"},"dialect":{"lastupdated":"2017-12-18","source":"1","classification":"C","value":"HK"},"occupation":{"lastupdated":"2017-12-18","source":"2","classification":"C","value":"21431","desc":"ENVIRONMENTAL ENGINEER"},"regadd":{"country":"SG","unit":"4","street":"BRADDELL HILL","lastupdated":"2017-12-19","block":"10G","source":"1","postal":"579726","classification":"C","floor":"5","building":"BRADDELL VIEW"},"mobileno":{"code":"","prefix":"","lastupdated":"","source":"4","classification":"C","nbr":""},"mailadd":{"country":"SG","unit":"15","street":"FULTON ROAD","lastupdated":"2017-12-19","block":"62","source":"2","postal":"578945","classification":"C","floor":"17","building":"FULTON HILL"},"billadd":{"country":"SG","unit":"4","street":"BALMORAL PARK","lastupdated":"2017-12-19","block":"25","source":"2","postal":"259854","classification":"C","floor":"2","building":"PINEWOOD GARDENS"},"hanyupinyinname":{"lastupdated":"","source":"1","classification":"C","value":""},"passportexpirydate":{"lastupdated":"","source":"1","classification":"C","value":""},"marriedname":{"lastupdated":"","source":"1","classification":"C","value":""},"relationships":[{"passportno":"","name":"","lastupdated":"","source":"2","classification":"C","type":"","idno":""}],"divorcedate":{"lastupdated":"","source":"1","classification":"C","value":""},"gradyear":{"lastupdated":"2017-12-19","source":"2","classification":"C","value":"1992"},"marriagecertno":{"lastupdated":"","source":"1","classification":"C","value":""},"householdincome":{"high":"","low":"","lastupdated":"","source":"2","classification":"C"},"email":{"lastupdated":"","source":"4","classification":"C","value":""},"birthcountry":{"lastupdated":"2017-12-18","source":"1","classification":"C","value":"SG"},"race":{"lastupdated":"2017-12-18","source":"1","classification":"C","value":"CN"},"housingtype":{"lastupdated":"2017-12-18","source":"1","classification":"C","value":"139"},"sex":{"lastupdated":"2017-12-18","source":"1","classification":"C","value":"F"},"countryofmarriage":{"lastupdated":"","source":"1","classification":"C","value":""},"hdbtype":{"lastupdated":"","source":"1","classification":"C","value":""},"marriagedate":{"lastupdated":"","source":"1","classification":"C","value":""},"employment":{"lastupdated":"2017-12-18","source":"2","classification":"C","value":"SMRT CORP PTD LTD"},"residentialstatus":{"lastupdated":"","source":"1","classification":"C","value":""},"passportnumber":{"lastupdated":"","source":"1","classification":"C","value":""},"schoolname":{"lastupdated":"2017-12-19","source":"2","classification":"C","value":"TP","desc":"TEMASEK POLYTECHNIC"},"hanyupinyinaliasname":{"lastupdated":"","source":"1","classification":"C","value":""},"marital":{"lastupdated":"2017-12-18","source":"1","classification":"C","value":"1"},"aliasname":{"lastupdated":"","source":"1","classification":"C","value":""},"childrenbirthrecords":[{"dialect":"MY","race":"MY","tob":"0625","sex":"M","source":"1","classification":"C","birthcertno":"T1132123J","hanyupinyinname":"","hanyupinyinaliasname":"","marriedname":"","aliasname":"","dob":"2011-12-31","name":"AYRIL ARYAN BIN SAIFUL RIZAL","lastupdated":"2018-05-07","secondaryrace":""}],"nationality":{"lastupdated":"2017-12-18","source":"1","classification":"C","value":"SG"},"vehno":{"lastupdated":"","source":"2","classification":"C","value":""},"dob":{"lastupdated":"2017-12-18","source":"1","classification":"C","value":"1974-05-11"},"name":{"lastupdated":"2017-12-18","source":"1","classification":"C","value":"LEE HUI LING"},"secondaryrace":{"lastupdated":"","source":"1","classification":"C","value":""}}
